+++ src/com/intellij/idea/plugin/hybris/project/HybrisProjectIconProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -15,19 +15,21 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.intellij.idea.plugin.hybris.codeInsight.hints
+package com.intellij.idea.plugin.hybris.project
 
-import com.intellij.codeInsight.hints.declarative.InlayHintsProvider
-import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent
-import com.intellij.openapi.editor.Editor
-import com.intellij.psi.PsiFile
+import com.intellij.ide.IconProvider
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.psi.PsiElement
+import javax.swing.Icon
 
-class DynamicAttributeDeclarativeInlayProvider : InlayHintsProvider {
+class HybrisProjectIconProvider : IconProvider() {
 
-    private val collector by lazy {
-        DynamicAttributeDeclarativeInlayHintsCollector()
+    override fun getIcon(p0: PsiElement, p1: Int): Icon? {
+        val file = p0.containingFile ?: return null
+        if (file.name.endsWith(HybrisConstants.IMPORT_OVERRIDE_FILENAME)) {
+            return HybrisIcons.PLUGIN_SETTINGS
+        }
+        return null
     }
-
-    override fun createCollector(file: PsiFile, editor: Editor) = if (HybrisProjectSettingsComponent.getInstance(file.project).isHybrisProject()) collector
-    else null
 }